src/requirements.txt
numpy
opencv-python
scikit-learn
matplotlib
kagglehub
python-dotenv